# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  "/api/v1/public/auth/register":
    post:
      tags:
        - Public Authentication Management (no token needed)
      summary: Register a new user.
      description: Creates a new client account in the system with the provided credentials.
        This endpoint is public.
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RegisterRequestDto"
        required: true
      responses:
        '201':
          description: User successfully registered.
        '400':
          description: Invalid registration data or username already exists.
  "/api/v1/public/auth/refresh":
    post:
      tags:
        - Public Authentication Management (no token needed)
      summary: Refresh access token.
      description: Validates the provided HttpOnly refresh token cookie. Issues a
        new access token and updates the refresh token cookie. This endpoint is public.
      operationId: refresh
      parameters:
        - name: refreshToken
          in: cookie
          description: The secure HttpOnly refresh token required to get a new access
            token.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tokens successfully refreshed.
          headers:
            Set-Cookie:
              description: Contains the updated HttpOnly 'refreshToken' cookie.
              style: simple
              schema:
                type: string
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TokenResponseDto"
        '401':
          description: Refresh token is expired, invalid, or missing.
...
  "/api/v1/public/auth/admin/register":
    post:
      tags:
        - Public Authentication Management (no token needed)
      summary: Complete administrative registration.
      description: Registers an administrative user account using a special sign-up
        token or pre-authorized invitation payload. This endpoint is public.
      operationId: completeAdminRegistration
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AdminRegisterDto"
        required: true
      responses:
        '200':
          description: Administrative user successfully registered.
        '400':
          description: Invalid administrative registration payload.
        '403':
          description: The registration token or signature is invalid.
...
  "/api/v1/private/auth/admin/invitations":
    post:
      tags:
        - Private Authentication Management
      summary: Create an administrative invitation.
      description: Generates a unique registration token or invitation link sent to
        the specified email to allow a new admin user to register. Requires **ROLE_ADMIN**
        authority.
      operationId: inviteNewAdmin
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AdminInviteRequestDto"
        required: true
      responses:
        '200':
          description: Administrative invitation successfully created.
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/AdminInvitationResponseDto"
        '400':
          description: Invalid invitation request payload or email format.
        '401':
          description: User is not authenticated.
        '403':
          description: User does not have the required administrative permissions.
      security:
        - BearerAuth: []
  "/api/v1/admin/users/search":
    post:
      tags:
        - Admin Users Management
      summary: Search users with pagination. (YES, it's POST)
      description: Returns a paginated list of users matching the search criteria
        (exact match for username, partial match for card holder name). Requires **ROLE_ADMIN**
        authority.
      operationId: searchUsers
      parameters:
        - name: pageable
          in: query
          description: Pagination parameters including page, size, and sort.
          required: true
          schema:
            type: object
          example:
            page: 0
            size: 10
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/GetUsersRequestDto"
        required: true
      responses:
        '200':
          description: User list successfully retrieved.
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/GetUsersResponseDto"
        '400':
          description: Invalid search or pagination parameters provided.
        '401':
          description: User is not authenticated.
        '403':
          description: User does not have the required administrative permissions.
      security:
        - BearerAuth: []
  "/api/v1/admin/cards":
    post:
      tags:
        - Admin Cards Management
      summary: Create a new bank card.
      description: Issues a new bank card based on the provided request details. Requires
        **ROLE_ADMIN** authority.
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateCardRequestDto"
        required: true
      responses:
        '201':
          description: Card successfully created.
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CreateCardResponseDto"
        '400':
          description: Invalid payload or validation error occurred.
        '401':
          description: User is not authenticated.
        '403':
          description: User does not have the required administrative permissions.
      security:
        - BearerAuth: []
    delete:
      tags:
        - Admin Cards Management
      summary: Delete a bank card.
      description: Permanently deletes or soft-deletes a specific bank card from the
        system. Requires **ROLE_ADMIN** authority.
      operationId: deleteCard
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DeleteCardRequestDto"
        required: true
      responses:
        '200':
          description: Card successfully deleted.
        '400':
          description: Invalid target card parameters provided.
        '401':
          description: User is not authenticated.
        '403':
          description: User does not have the required administrative permissions.
        '404':
          description: Target bank card not found.
      security:
        - BearerAuth: []
  "/api/v1/admin/cards/search":
    post:
      tags:
        - Admin Cards Management
      summary: Get filtered list of all cards. (YES, it's POST)
      description: Returns a paginated list of bank cards matching the search criteria
        specified in the request body. Requires **ROLE_ADMIN** authority.
      operationId: getAllCards
      parameters:
        - name: pageable
          in: query
          description: Pagination parameters including page, size, and sort.
          required: true
          schema:
            type: object
          example:
            page: 0
            size: 10
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/GetCardsAdminRequestDto"
        required: true
      responses:
        '200':
          description: Card list successfully retrieved.
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/GetCardsResponseDto"
        '400':
          description: Invalid search or pagination parameters provided.
        '401':
          description: User is not authenticated.
        '403':
          description: User does not have the required administrative permissions.
      security:
        - BearerAuth: []
  "/api/v1/admin/cards/status":
    patch:
      tags:
        - Admin Cards Management
      summary: Update card status.
      description: Changes the status of a specific bank card (e.g., active, blocked,
        suspended). Requires **ROLE_ADMIN** authority.
      operationId: updateStatus
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateCardStatusRequestDto"
        required: true
      responses:
        '200':
          description: Card status successfully updated.
        '400':
          description: Invalid status value or missing payload fields.
        '401':
          description: User is not authenticated.
        '403':
          description: User does not have the required administrative permissions.
        '404':
          description: Target bank card not found.
      security:
        - BearerAuth: []
components:
  schemas:
...
    GetCardsResponseDto:
      type: object
      description: Response payload containing the paginated collection of retrieved
        bank cards along with execution metadata.
      properties:
        cards:
          type: array
          description: The list of bank card details matching the requested filter
            criteria or belonging to the target user.
          items:
            "$ref": "#/components/schemas/GetCardResponseDto"
        pageNumber:
          type: integer
          format: int32
          description: The current page index position within the paginated record
            set, starting from index 0.
          example: 0
        pageSize:
          type: integer
          format: int32
          description: The maximum number of card records requested or allowed to
            be displayed on a single page slice.
          example: 10
        totalElements:
          type: integer
          format: int64
          description: The absolute total count of all bank card records existing
            in the database that match the criteria.
          example: 47
        totalPages:
          type: integer
          format: int32
          description: The calculated total count of available pages based on total
            elements and the requested page size.
          example: 5
...
    RegisterRequestDto:
      type: object
      description: Request payload used by new consumers to sign up and create a bank
        user account.
      properties:
        userName:
          type: string
          description: Unique username chosen by the user for account identification.
          example: bank_customer_7
          maxLength: 50
          minLength: 4
        password:
          type: string
          description: Secure password created by the user for account access protection.
          example: P@ssword2026
          maxLength: 100
          minLength: 6
        email:
          type: string
          format: email
          description: Primary email address used for notifications and account verification.
          example: customer7@example.com
          minLength: 1
      required:
        - email
        - password
        - userName
    TokenResponseDto:
      type: object
      description: Response payload containing the generated authentication access
        token upon successful authentication.
      properties:
        accessToken:
          type: string
          description: The JWT access token used to authenticate subsequent protected
            resource requests.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        userName:
          type: string
          description: The unique username of the authenticated account holder.
          example: bank_customer_7
...
    AdminRegisterDto:
      type: object
      description: Request payload for completing the administrator registration using
        an invitation token.
      properties:
        userName:
          type: string
          description: Unique username for the new administrator account.
          example: admin_john
          maxLength: 50
          minLength: 4
        password:
          type: string
          description: Secure password for the administrator account.
          example: bAnk$ecure2026
          maxLength: 100
          minLength: 8
        token:
          type: string
          format: uuid
          description: The unique UUID invitation token sent via email.
          example: 123e4567-e89b-12d3-a456-426614174000
      required:
        - password
        - token
        - userName
    AdminInviteRequestDto:
      type: object
      description: Request payload used by other administrators to create invitation
        token.
      properties:
        email:
          type: string
          format: email
          description: The email where the administrative registration link/token
            will be sent.
          example: new_admin@bank.com
          minLength: 1
      required:
        - email
    AdminInvitationResponseDto:
      type: object
      description: Response payload with the details of a generated registration administrator
        token.
      properties:
        email:
          type: string
          format: email
          description: The email address to which the invitation token must be sent.
          example: new_admin@bank.com
        token:
          type: string
          format: uuid
          description: The unique registration token generated for the new administrator.
          example: 123e4567-e89b-12d3-a456-426614174000
        expiryDate:
          type: string
          format: date-time
          description: The exact timestamp when the invitation token will expire.
          example: '2026-08-14T16:33:00Z'
    GetUsersRequestDto:
      type: object
      description: Request payload used to search system users based on filtering
        criteria.
      properties:
        userName:
          type:
            - string
            - 'null'
          description: The exact authentication handle or login name of the target
            user. This field is optional.
          example: super_admin
        holderName:
          type:
            - string
            - 'null'
          description: The name of the cardholder linked to the bank cards. Supports
            partial or substring matches. This field is optional.
          example: JOHN DOE
    GetUserResponseDto:
      type: object
      description: Response payload containing the details of a specific user account.
      properties:
        uuid:
          type: string
          format: uuid
          description: The unique business identifier (UUID) assigned to the user
            account.
          example: e4219747-f42d-4e32-92a8-e67d9841910e
        userName:
          type: string
          description: The unique authentication handle and login name of the user.
          example: super_admin
        email:
          type: string
          description: The electronic mail address linked directly to the user profile.
          example: root@bank.com
        enabled:
          type: boolean
          description: The operational state indicating whether the user account is
            active or disabled.
          example: true
        role:
          type: string
          description: The administrative authority level and security role assigned
            to the user.
          enum:
            - USER
            - ADMIN
          example: ADMIN
    GetUsersResponseDto:
      type: object
      description: Response payload containing the paginated collection of system
        users.
      properties:
        users:
          type: array
          description: The list of users matching the specified search filtering criteria.
          items:
            "$ref": "#/components/schemas/GetUserResponseDto"
        pageNumber:
          type: integer
          format: int32
          description: The current page index position within the record set, starting
            from 0.
          example: 0
        pageSize:
          type: integer
          format: int32
          description: The maximum number of user records displayed on a single page
            slice.
          example: 10
        totalElements:
          type: integer
          format: int64
          description: The absolute total count of user records matching the filter
            criteria in the database.
          example: 25
        totalPages:
          type: integer
          format: int32
          description: The calculated total count of available pages based on the
            current page size.
          example: 3
    CreateCardRequestDto:
      type: object
      description: Request payload used to issue and provision a new bank card for
        a specific user.
      properties:
        userId:
          type: string
          format: uuid
          description: The unique identifier of the user for whom the card is being
            created.
          example: 123e4567-e89b-12d3-a456-426614174000
        holderName:
          type: string
          description: The full name of the cardholder, typically written in capital
            Latin letters.
          example: JOHN DOE
          minLength: 1
        initBalance:
          type: number
      required:
        - holderName
        - userId
    CreateCardResponseDto:
      type: object
      description: Response payload containing the details of a newly issued bank
        card.
      properties:
        cardNumber:
          type: string
          description: The masked credit/debit card number generated by the system.
          example: 4312 8900 1122 3344
        uuid:
          type: string
          format: uuid
          description: The unique identifier of the card.
          example: 123e4567-e89b-12d3-a456-426614174000
        holderName:
          type: string
          description: The full name of the cardholder in capital Latin letters.
          example: JOHN DOE
        expiryDate:
          type: string
          format: date-time
          description: The calculated expiration timestamp for the issued bank card.
          example: '2031-08-13T16:50:00Z'
        status:
          type: string
          description: The initial operational status assigned to the new card.
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: ACTIVE
        balance:
          type: number
          description: The initial financial balance allocated to the card upon creation.
          example: 0
    GetCardsAdminRequestDto:
      type: object
      description: Request payload used to fetch and filter a list of bank cards belonging
        to a specific user.
      properties:
        userId:
          type:
            - string
            - 'null'
          format: uuid
          description: The unique identifier of the user who owns the card. This field
            is optional.
          example: 123e4567-e89b-12d3-a456-426614174000
        status:
          type:
            - string
            - 'null'
          description: Optional card status filter. If omitted or null, all cards
            for the user will be returned.
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: ACTIVE
    UpdateCardStatusRequestDto:
      type: object
      description: Request payload used by administrators to update the operational
        status of a specific bank card.
      properties:
        cardId:
          type: string
          format: uuid
          description: The unique identifier of the card.
          example: 123e4567-e89b-12d3-a456-426614174000
        cardStatus:
          type: string
          description: The new status to be assigned to the bank card.
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: BLOCKED
      required:
        - cardId
        - cardStatus
    DeleteCardRequestDto:
      type: object
      description: Request payload used to permanently or logically delete a specific
        bank card from the system.
      properties:
        cardId:
          type: string
          format: uuid
          description: The unique identifier of the bank card that needs to be deleted.
          example: 987f6543-e21b-78d3-c987-123456789abc
      required:
        - cardId